message from the bottle

+++ random-book-from-db.py
@@ -7,7 +7,7 @@ from decimal import Decimal
 class DecimalEncoder(json.JSONEncoder):
     def default(self, o):
         if isinstance(o, Decimal):
-            return float(o)
+            return int(o)
         return super(DecimalEncoder, self).default(o)
 
 # Create a connection to DynamoDB Local
